+++ app/src/main/java/be/ugent/zeus/hydra/service/urgent/MediaNotificationBuilder.java
@@ -1,12 +1,10 @@
-package be.ugent.zeus.hydra.service.urgent.media;
+package be.ugent.zeus.hydra.service.urgent;
 
 import android.app.Notification;
 import android.app.PendingIntent;
 import android.content.Context;
 import android.content.Intent;
 import android.graphics.BitmapFactory;
-import android.support.annotation.DrawableRes;
-import android.support.annotation.StringRes;
 import android.support.v4.app.NotificationCompat;
 import android.support.v4.media.session.MediaButtonReceiver;
 import android.support.v4.media.session.MediaSessionCompat;
@@ -14,8 +12,6 @@ import android.support.v4.media.session.PlaybackStateCompat;
 
 import be.ugent.zeus.hydra.R;
 import be.ugent.zeus.hydra.data.ChannelCreator;
-import be.ugent.zeus.hydra.service.urgent.MusicService;
-import be.ugent.zeus.hydra.service.urgent.track.Track;
 import be.ugent.zeus.hydra.ui.main.MainActivity;
 
 /**
@@ -24,8 +20,6 @@ import be.ugent.zeus.hydra.ui.main.MainActivity;
  * @author Niko Strijbol
  */
 public class MediaNotificationBuilder {
-
-    private static final int NOTIFICATION_ID = 1;
 
     private final Context context;
 
@@ -94,14 +88,6 @@ public class MediaNotificationBuilder {
         return builder.build();
     }
 
-    private NotificationCompat.Action generateAction(@DrawableRes int icon, @StringRes int title, String intentAction) {
-        Intent intent = new Intent(context, MusicService.class);
-        intent.setAction(intentAction);
-        PendingIntent pendingIntent = PendingIntent.getService(context, NOTIFICATION_ID, intent, PendingIntent.FLAG_UPDATE_CURRENT);
-        String titleString = context.getString(title);
-        return new NotificationCompat.Action.Builder(icon, titleString, pendingIntent).build();
-    }
-
     public interface MediaInfoProvider {
 
         boolean isPlaying();
@@ -117,21 +103,5 @@ public class MediaNotificationBuilder {
         Intent startThis = new Intent(context, MainActivity.class);
         startThis.putExtra(MainActivity.ARG_TAB, R.id.drawer_urgent);
         return PendingIntent.getActivity(context, 0, startThis, PendingIntent.FLAG_UPDATE_CURRENT);
-    }
-
-    public Notification buildPreparingNotification() {
-
-        // Get the click intent
-        PendingIntent clickIntent = buildClickIntent();
-
-        return new NotificationCompat.Builder(context, ChannelCreator.URGENT_CHANNEL)
-                .setSmallIcon(R.drawable.ic_notification_urgent)
-                .setShowWhen(false)
-                .setContentTitle("TEST PREPARING")
-                .setContentIntent(clickIntent)
-                .setVisibility(NotificationCompat.VISIBILITY_PUBLIC)
-                .setChannelId(ChannelCreator.URGENT_CHANNEL)
-                //.setLargeIcon(BitmapFactory.decodeResource(context.getResources())VectorDrawableCompat.create(context.getResources(), R.drawable.ic_urgent_notification, null).)
-                .build();
     }
 }
