+++ app/src/main/java/be/ugent/zeus/hydra/service/urgent/Track.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package be.ugent.zeus.hydra.service.urgent.track;
+package be.ugent.zeus.hydra.service.urgent;
 
 import android.graphics.Bitmap;
 import android.support.annotation.NonNull;
